Fixed a few issues and interval tests now pass.

+++ hyracks-fullstack/hyracks/hyracks-dataflow-std/src/main/java/org/apache/hyracks/dataflow/std/join/RunFileStream.java
@@ -181,7 +181,9 @@ public class RunFileStream {
     }
 
     public void removeRunFile() {
-        FileUtils.deleteQuietly(runfile.getFile());
+        if (runfile != null) {
+            FileUtils.deleteQuietly(runfile.getFile());
+        }
     }
 
     public long getReadPointer() {
